Items vehicle collision (#546)
* Work

* Fix garbage item entries not cleared after level reload

* Fix collisions for items, add shatterable meshes

* Convert GetLaraInfo to struct method

* Unify all vehicle item collision code

* Update motorbike.cpp

* Add UpperCeilingBound setup which indirectly affects vehicle coll results

* Cleanup vehicle collision code, add checks for all vehicles

* Add condition for undead enemies

* Revert GetLaraInfo conversion

* Update items.cpp

* Reduce vehicle radius a little

* Remove shady ternary operation

* Update minecart.cpp

* Fix small things

* Don't kill baddies if vehicle speed is too low

* Fix infinite loop

* Add jeep shallow water functionality

* Generalize vehicle explosion, make all vehicles use shallow water, add explosion depth tolerance

* Fix splashes

* Update collide_item.cpp

* Revert "Update collide_item.cpp"

This reverts commit 078bc5b7620b580494aa883bca14f78cd68198ad.

+++ TombEngine/Objects/TR4/Vehicles/jeep_info.h
@@ -8,7 +8,7 @@ struct JeepInfo {
 	int revs;
 	short engineRevs;
 	short trackMesh;
-	int jeepTurn;
+	short jeepTurn;
 	int fallSpeed;
 	short momentumAngle;
 	short extraRotation;
